[RLlib] Issue 9071 A3C w/ RNN not working due to VF assuming no RNN. (#13238)
