+++ rllib/agents/dqn/dqn_tf_policy.py
@@ -301,17 +301,11 @@ def adam_optimizer(policy: Policy, config: TrainerConfigDict
 
 def clip_gradients(policy: Policy, optimizer: "tf.keras.optimizers.Optimizer",
                    loss: TensorType) -> ModelGradients:
-    if policy.config["grad_clip"] is not None:
-        grads_and_vars = minimize_and_clip(
-            optimizer,
-            loss,
-            var_list=policy.q_func_vars,
-            clip_val=policy.config["grad_clip"])
-    else:
-        grads_and_vars = optimizer.compute_gradients(
-            loss, var_list=policy.q_func_vars)
-    grads_and_vars = [(g, v) for (g, v) in grads_and_vars if g is not None]
-    return grads_and_vars
+    return minimize_and_clip(
+        optimizer,
+        loss,
+        var_list=policy.q_func_vars,
+        clip_val=policy.config["grad_clip"])
 
 
 def build_q_stats(policy: Policy, batch) -> Dict[str, TensorType]:
