+++ rllib/agents/dreamer/dreamer_torch_policy.py
@@ -1,11 +1,11 @@
 import logging
 
 import ray
-from ray.rllib.agents.a3c.a3c_torch_policy import apply_grad_clipping
 from ray.rllib.agents.dreamer.utils import FreezeParameters
 from ray.rllib.models.catalog import ModelCatalog
 from ray.rllib.policy.policy_template import build_policy_class
 from ray.rllib.utils.framework import try_import_torch
+from ray.rllib.utils.torch_ops import apply_grad_clipping
 
 torch, nn = try_import_torch()
 if torch:
