+++ rllib/agents/impala/impala.py
@@ -1,7 +1,6 @@
 import logging
 
 import ray
-from ray.rllib.agents.a3c.a3c_tf_policy import A3CTFPolicy
 from ray.rllib.agents.impala.vtrace_tf_policy import VTraceTFPolicy
 from ray.rllib.agents.trainer import Trainer, with_common_config
 from ray.rllib.agents.trainer_template import build_trainer
@@ -160,6 +159,7 @@ def get_policy_class(config):
         if config["vtrace"]:
             return VTraceTFPolicy
         else:
+            from ray.rllib.agents.a3c.a3c_tf_policy import A3CTFPolicy
             return A3CTFPolicy
